requirements.txt
Django==3.2.7
dataclasses
dnspython
djangorestframework==3.12.4
pymongo==3.10.1
PyYAML==5.4.1
black==21.9b0 
